backend/requirements.txt
boto3==1.28.85
botocore==1.31.85
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
pytest==7.4.3
pytest-asyncio==0.21.1
flask==3.0.0
flask-cors==4.0.0

